requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
captcha>=0.4
numpy>=1.20.0
matplotlib>=3.3.0
tqdm>=4.60.0